Goods 굿즈 / Bid  경매 기능 통합 / Wallet  지갑 / Comment  댓글 기능 통합 PR

+++ backend/src/main/java/com/ssafy/backend/user/service/UserServiceImpl.java
@@ -1,13 +1,18 @@
 package com.ssafy.backend.user.service;
 
 import com.ssafy.backend.common.exception.CustomException;
+import com.ssafy.backend.goods.model.GoodsMapper;
+import com.ssafy.backend.goods.model.GoodsResponseDto;
 import com.ssafy.backend.user.model.*;
+import com.ssafy.backend.wallet.model.WalletMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Collections;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -15,6 +20,8 @@ import java.util.List;
 public class UserServiceImpl implements UserService {
 
     private final UserMapper userMapper;
+    private final GoodsMapper goodsMapper;
+    private final WalletMapper walletMapper;
 
     /**
      * 내 프로필 조회
@@ -63,7 +70,9 @@ public class UserServiceImpl implements UserService {
                 .likedAnimeId3(request.getLikedAnimeId3())
                 .build();
 
-        userMapper.updateUser(updatePayload);
+        int updateUser = userMapper.updateUser(updatePayload);
+        if(updateUser < 1) throw new CustomException("프로필 수정에 실패하였습니다.", HttpStatus.INTERNAL_SERVER_ERROR);
+
 
         List<AnimeSelectionDto> likedAnimes = userMapper.findUserLikedAnimes(userId);
 
@@ -92,6 +101,43 @@ public class UserServiceImpl implements UserService {
     @Override
     @Transactional
     public void deleteMyAccount(Long userId) {
-        userMapper.deleteUser(userId);
+
+        // 1. 등록한 굿즈가 있을 경우
+        boolean existGoodsResult = goodsMapper.existsActiveGoodsByUserId(userId);
+        //1-1. 'PROCEEDING'(진행중) 인 경우 삭제 불가
+        if(existGoodsResult) {
+            throw new CustomException("진행 중인 경매가 있어 회원 탈퇴를 할 수 없습니다.", HttpStatus.BAD_REQUEST);
+        }
+
+        //1-2. 모든 굿즈 글이 'PROCEEDING'(진행중) 이 아닌 경우 삭제
+        int deleteResult = goodsMapper.deleteGoods(null, userId);
+        if(deleteResult < 1) throw new CustomException("굿즈 글 삭제에 실패하였습니다.", HttpStatus.INTERNAL_SERVER_ERROR);
+
+        // 2. 예치금이 0원일 경우에만 삭제 가능
+        int lockedBalance = walletMapper.selectCoinWallet(userId).getLockedBalance();
+        if(lockedBalance != 0) throw new CustomException("진행 중인 경매가 있어 회원 탈퇴를 할 수 없습니다.", HttpStatus.BAD_REQUEST);
+
+        int deleteUser = userMapper.deleteUser(userId);
+        if(deleteUser < 1) throw new CustomException("회원 탈퇴에 실패하였습니다.", HttpStatus.INTERNAL_SERVER_ERROR);
+    }
+
+    /**
+     * 마이페이지 내 사용자가 등록한 굿즈 목록 조회
+     */
+    @Override
+    public List<GoodsResponseDto.MyPageInRegisteredGoodsCard> getAllRegisteredGoods(Long loginUserId) {
+        // 데이터 조회 후, 없으면 빈 리스트 던지기
+        return Optional.ofNullable(goodsMapper.selectAllRegisteredGoods(loginUserId))
+                .orElse(Collections.emptyList());
+    }
+
+    /**
+     * 마이페이지 내 사용자가 참여한 굿즈 목록 조회
+     */
+    @Override
+    public List<GoodsResponseDto.MyPageInParticipatedGoodsCard> getAllParticipatedGoods(Long loginUserId) {
+        // 데이터 조회 후, 없으면 빈 리스트 던지기
+        return Optional.ofNullable(goodsMapper.selectAllParticipatedGoods(loginUserId))
+                .orElse(Collections.emptyList());
     }
 }
